remove deprecated dtype getters from torch.testing
[ghstack-poisoned]
